SSOR-61 Refactored to use boss-core submodule. Added swagger.
