+++ .gitmodules
@@ -1,0 +1,4 @@
+[submodule "boss-core"]
+	path = boss-core
+	url = https://github.com/SSOrphans/boss-core.git
+	branch = refactor/SSOR-165-Card_Submodule
