+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.url=jdbc:mysql://localhost:3306/boss
 spring.datasource.username=root
 spring.datasource.password=password
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-#spring.jpa.database-platform=org.hibernate.dialect.MySQL5InnoDBDialect
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
